Moved the split transformation into its own module.
